+++ src/Page/ChangePasswordPage/style.css.ts
@@ -1,6 +1,4 @@
 import { style } from '@vanilla-extract/css';
-import baseFontStyle from '../../styles/font.css';
-import { themeColors } from '../../styles/theme.css';
 
 export const pageStyle = style({
   display: 'flex',
@@ -11,11 +9,3 @@ export const pageStyle = style({
 
   padding: '4.375rem',
 });
-
-export const titleStyle = style([
-  baseFontStyle.title,
-  {
-    textAlign: 'center',
-    color: themeColors.text[1],
-  },
-]);
